rename fskit
