+++ fskit/file.go
@@ -1,4 +1,4 @@
-package fs
+package fskit
 
 import (
 	"errors"
